PARAM_HAVE_WRAPPER_ALL
git-svn-id: file:///Users/walter/linux/walter/src/svnroot/trunk/bbs100@196 d90a1141-1c03-0410-82d0-813645769019

+++ src/ConnData.c
@@ -141,7 +141,7 @@ socklen_t client_len = sizeof(struct sockaddr_storage);
 	}
 	strcpy(new_conn->hostname, new_conn->ipnum);
 
-	if (!allow_Wrapper(new_conn->ipnum, WRAPPER_ALL_USERS)) {
+	if (PARAM_HAVE_WRAPPER_ALL && !allow_Wrapper(new_conn->ipnum, WRAPPER_ALL_USERS)) {
 		write_Conn(new_conn, "error locked out\n");
 		log_auth("connection from %s closed by wrapper", new_conn->ipnum);
 		destroy_Conn(new_conn);
